+++ NavigationReact/src/HistoryCacheContext.ts
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export default createContext({});
