+++ NavigationReact/src/NavigationHandler.tsx
@@ -1,8 +1,9 @@
 'use client'
-import React, { useCallback, useDeferredValue, useEffect, useMemo, useState, useTransition } from 'react';
+import React, { useCallback, useDeferredValue, useEffect, useMemo, useRef, useState, useTransition } from 'react';
 import { StateNavigator, StateContext, State } from 'navigation';
 import NavigationContext from './NavigationContext.js';
 import RefetchContext from './RefetchContext.js';
+import HistoryCacheContext from './HistoryCacheContext.js';
 import NavigationDeferredContext from './NavigationDeferredContext.js';
 type NavigationHandlerState = { ignoreCache?: boolean, oldState: State, state: State, data: any, asyncData: any, stateNavigator: StateNavigator };
 
@@ -10,6 +11,7 @@ const NavigationHandler = ({ stateNavigator, children }: { stateNavigator: State
     const [navigationEvent, setNavigationEvent] = useState<{ data: NavigationHandlerState, stateNavigator: StateNavigator, resumeNavigation?: () => void }>();
     const navigationDeferredEvent = useDeferredValue(navigationEvent);
     const [isPending, startTransition] = useTransition();
+    const historyCacheRef = useRef({});
     const raiseNavigationEvent = useCallback((stateContext: StateContext = stateNavigator.stateContext, resumeNavigation?: () => void) => {
         class AsyncStateNavigator extends StateNavigator {
             constructor() {
@@ -59,9 +61,26 @@ const NavigationHandler = ({ stateNavigator, children }: { stateNavigator: State
         stateNavigator.onNavigate(onNavigate);
         return () => stateNavigator.offNavigate(onNavigate);
     }, [stateNavigator, navigationEvent, raiseNavigationEvent]);
+    const oldSceneCount = (typeof window !== 'undefined' && window.history.state?.sceneCount) || 0;
     useEffect(() => {
-        if (!isPending && navigationEvent === navigationDeferredEvent)
+        if (!isPending && navigationEvent === navigationDeferredEvent) {
             navigationEvent.resumeNavigation?.();
+            const { stateContext: { nextCrumb, historyAction, history } } = navigationEvent.stateNavigator;
+            const url = nextCrumb?.crumblessUrl;
+            if (historyAction === 'none') return;
+            const historyCache = historyCacheRef.current;
+            const sceneCount = window.history.state?.sceneCount || (oldSceneCount + 1);
+            if (!historyCache[url]) historyCache[url] = {};
+            historyCache[url].count = Math.min(historyCache[url].count || sceneCount, sceneCount);
+            const historyUrls = Object.keys(historyCache);
+            for(let i = 0; i < historyUrls.length && !history; i++) {
+                const historyUrl = historyUrls[i];
+                const gap = historyCache[historyUrl].count - sceneCount;
+                if (historyUrl !== url && (gap === 0 || (historyAction === 'add' && gap > 0)))
+                    delete historyCache[historyUrl];
+            }
+            window.history.replaceState({...window.history.state, sceneCount}, null);
+        }
     }, [isPending, navigationEvent, navigationDeferredEvent]);
     useEffect(() => {
         if (stateNavigator !== navigationEvent.stateNavigator)
@@ -69,11 +88,13 @@ const NavigationHandler = ({ stateNavigator, children }: { stateNavigator: State
     }, [navigationEvent, stateNavigator])
     return (
         <NavigationContext.Provider value={navigationEvent?.data}>
-            <RefetchContext.Provider value={refetchControl}>
-                <NavigationDeferredContext.Provider value={navigationDeferredEvent?.data}>
-                    {children}
-                </NavigationDeferredContext.Provider>
-            </RefetchContext.Provider>
+            <NavigationDeferredContext.Provider value={navigationDeferredEvent?.data}>
+                <RefetchContext.Provider value={refetchControl}>
+                    <HistoryCacheContext.Provider value={historyCacheRef.current}>
+                        {children}
+                    </HistoryCacheContext.Provider>
+                </RefetchContext.Provider>
+            </NavigationDeferredContext.Provider>
         </NavigationContext.Provider>
     )
 }
